+++ assets/javascripts/kitten/components/graphics/animations/loader-with-paragraph/stories.js
@@ -49,4 +49,6 @@ export default {
 
 export const Default = args => <LoaderWithParagraph {...args} />
 
-export const WithCustomComponent = args => <LoaderWithParagraph {...args} loaderComponent={KissKissLoadingAnimation} />
+export const WithCustomComponent = args => (
+  <LoaderWithParagraph {...args} loaderComponent={KissKissLoadingAnimation} />
+)